% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyslider.R
\name{easyslider}
\alias{easyslider}
\alias{sliderFilter}
\alias{slider2Filter}
\alias{dropdownFilter}
\title{Easy sliders for shiny.}
\usage{
sliderFilter(df, aes, ...)

slider2Filter(df, aes, ...)

dropdownFilter(df, aes, ...)
}
\description{
This package was written because it was frustrating to
add a slider in shiny, which typically required changing
code in three places.
}
