% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyslider.R
\name{sliderFilter}
\alias{sliderFilter}
\alias{slider2Filter}
\alias{dropdownFilter}
\alias{easySliderUIOutput}
\title{Easy slider functions}
\usage{
sliderFilter(df, aes, ...)

slider2Filter(df, aes, ...)

dropdownFilter(df, aes, ...)

easySliderUIOutput()
}
\arguments{
\item{df}{a data.frame}

\item{aes}{an aesthic to map a column to the filter}
}
\value{
a reactive, filtered data.frame
}
\description{
Easy slider functions
}
