% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyslider.R
\docType{package}
\name{easyslider}
\alias{easyslider}
\alias{easyslider-package}
\title{easyslider: Easy sliders for shiny.}
\description{
Quickly add UI components by aesthetic mapping
}
\details{
This package was written because it was frustrating to
add a slider in shiny, which typically required changing
code in three places: the placeholder in \code{ui.R},
a \code{renderUI} in \code{server.R} and also wiring it
up to a plot by using \code{input$thing}.

Instead, we can build simple UIs more conviniently by
generating the components from aesthetics. In your server
function, pipe data through some filters, then generate
plots and tables appropriately.
}
\examples{
\dontrun{

#ui.R

library(shiny)


shinyUI(fluidPage(

  # Application title
  titlePanel("Easy Slider Diamond Demo"),

  # Sidebar with easyslider controls
  sidebarLayout(
    sidebarPanel(
      easySliderUIOutput()
    ),

    # Show a plot of the filtered data
    mainPanel(
      plotOutput("distPlot")
    )
  )

))

#server.R

require(dplyr)
require(ggplot2)

library(shiny)
library(easyslider)


shinyServer(function(input, output) {

  df <- diamonds \%>\%
    slider2Filter(aes(depth)) \%>\%
    dropdownFilter(aes(clarity)) \%>\%
    checkboxGroupFilter(aes(cut))
    checkboxGroupFilter(aes(cut))

  output$distPlot <- renderPlot({
      df() \%>\% ggplot() + aes(x=carat, y=price, color=cut) + geom_point()
    })

})


}
}
