% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyslider.R
\name{sliderFilter}
\alias{sliderFilter}
\alias{slider2Filter}
\alias{dropdownFilter}
\alias{easySliderUIOutput}
\title{Easy slider functions}
\usage{
sliderFilter(df, aes, label, ...)

slider2Filter(df, aes, label, ...)

dropdownFilter(df, aes, label, ..., sfn = I, all.levels = FALSE)

easySliderUIOutput()
}
\arguments{
\item{df}{a data.frame}

\item{aes}{an aesthic to map a column to the filter}

\item{label}{the label to use on the filter}

\item{sfn}{a function that returns the dropdown items you want to display, given the unique values
present in the column specified by the aesthetic (eg sort, sample, etc)}

\item{all.levels}{Add an "(All Levels)" option.}
}
\value{
a reactive, filtered data.frame
}
\description{
Easy slider functions
}
